+++ src/commands/resourceCommand.ts
@@ -77,7 +77,7 @@ async function unautoRefreshRes() {
 async function copyGres() {
     const selected = resourceTreeView.selection[0];
     if (selected instanceof ListItem){
-        const gresString = (selected.title as string).replace(/\s*\(.*\)$/, '');
+        const gresString = selected.title.replace(/\s*\(.*\)$/, '');
         vscode.env.clipboard.writeText(gresString);
     } else if (selected instanceof NodeItem){
         vscode.env.clipboard.writeText(selected.node.nodeid);