+++ CHANGELOG.md
@@ -8,11 +8,15 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 
 ### Added
 
-- Add `submit time`, `start time` and `end time` in task info.
+- Add `submit time`, `start time` and `end time` in task info
 
 ### Changed
 
-- Change `save task` method to async for better performance.
+- Change `save task` method to async for better performance
+
+### Fixed
+
+- Fixed the node will not show correctly after running, if the task did not start immediately upon submission
 
 ## [0.7.2][0.7.2] - 2025-04-01