+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 
 - Add `submit time`, `start time` and `end time` in task info
 - Add `Set display information` button and command to custom display tasks information
+- Add cache for async functions to avoid repeated calls
 
 ### Changed
 
@@ -18,6 +19,7 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 ### Fixed
 
 - Fixed the node will not show correctly after running, if the task did not start immediately upon submission
+- Fixed the task marked as finished would be set to finished every refresh
 
 ## [0.7.2][0.7.2] - 2025-04-01